refactor: move pages into route groups and move home-view
