+++ app/_components/home/home-view.tsx
@@ -2,7 +2,7 @@
 
 import Container from "@/common/components/atoms/layouts/Container";
 import { Grid } from "@/common/components/atoms/layouts/Grid";
-import Hero from "@/app/home/_components/Hero/Hero";
+import Hero from "@/app/_components/home/Hero/Hero";
 import { getSearchRecommendations } from "@/lib/actions/recommendations";
 import { useState } from "react";
