+++ app/page.tsx
@@ -1,4 +1,4 @@
-import HomeView from "../common/views/home-view";
+import HomeView from "./_components/home/home-view";
 
 export default function Page() {
   return <HomeView />;
